+++ programs/super-liquidity/src/instructions/deposit.rs
@@ -32,8 +32,8 @@ impl<'info> Deposit<'info> {
         check_token_position(&self.global_state, &self.mint, position) && 
         check_vault(&self.user_account.key, &self.user_vault)
     )]
-    pub fn process(&mut self, amount: u64, position: usize) -> Result<()> {
-        let vault = &mut self.user_vault.vaults[position];
+    pub fn process(&mut self, amount: u64, position: u8) -> Result<()> {
+        let vault = &mut self.user_vault.vaults[position as usize];
 
         if self.get_token_from.amount < amount {
             msg!(
